Conflicts:
	community/consistency-check-legacy/src/main/java/org/neo4j/legacy/consistency/ConsistencyCheckService.java
	community/consistency-check/src/main/java/org/neo4j/consistency/ConsistencyCheckService.java
	community/kernel/src/main/java/org/neo4j/kernel/impl/store/DumpStore.java
	community/kernel/src/main/java/org/neo4j/kernel/impl/storemigration/StoreFile.java
	community/kernel/src/main/java/org/neo4j/kernel/impl/storemigration/legacystore/v21/propertydeduplication/PropertyDeduplicator.java
	community/kernel/src/main/java/org/neo4j/unsafe/batchinsert/BatchInserterImpl.java
	community/kernel/src/test/java/org/neo4j/kernel/impl/store/CommonAbstractStoreTest.java
	enterprise/backup/src/test/java/org/neo4j/backup/TestBackup.java

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/pagecache/ConfiguringPageCacheFactory.java
@@ -127,7 +127,7 @@ public class ConfiguringPageCacheFactory
     public void dumpConfiguration()
     {
         int cachePageSize = calculatePageSize( config, swapperFactory );
-        int maxPages = calculateMaxPages( config, cachePageSize );
+        long maxPages = calculateMaxPages( config, cachePageSize );
         long totalPhysicalMemory = totalPhysicalMemory();
         String totalPhysicalMemMb = totalPhysicalMemory == -1? "?" : "" + totalPhysicalMemory / 1024 / 1024;
         long maxVmUsageMb = Runtime.getRuntime().maxMemory() / 1024 / 1024;
